% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_use_fst.R
\name{cache_register_fst}
\alias{cache_register_fst}
\alias{cache_use_fst}
\title{Cache objects using fst (fst)}
\usage{
cache_register_fst()

cache_use_fst()
}
\description{
Enables caching using the \emph{fst} package
}
\details{
You probably should use \code{\link[=cache_use_fst]{cache_use_fst()}} instead which makes fst
the default method for reading/writing to the cache.  There is generally no
reason to invoke this function as it is called when the packages is
attached.
}
\examples{
dontrun{
 cache_use_fst()
 cache(mtcars)
 if( exists('mtcars') ) rm(mtcars)
 uncache(mtcars)
 fs::dir_ls( cache_find() )
}

}
\seealso{
\itemize{
\item \code{\link[=cache_use_fst]{cache_use_fst()}}
}
}
